Server/requirements.txt
alembic==1.4.1
appdirs==1.4.3
asgiref==3.2.3
bleak==0.6.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
filelock==3.0.12
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
paho-mqtt==1.5.0
pipenv==2018.11.26
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pythonnet==2.4.0
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.3
Werkzeug==1.0.0
whitenoise==5.0.1
zipp==3.1.0
